:construction: Update tasks and projects sections.
